Refactoring

See merge request !1

+++ beans/src/test/java/test/entities/TestConstants.java
@@ -2,21 +2,22 @@ package test.entities;
 
 import com.lapsa.insurance.elements.InsuranceClassType;
 import com.lapsa.insurance.elements.SubjectType;
-import com.lapsa.insurance.esbd.domain.entities.general.SubjectCompanyEntity;
-import com.lapsa.insurance.esbd.domain.entities.general.SubjectPersonEntity;
+
+import tech.lapsa.insurance.esbd.entities.SubjectCompanyEntity;
+import tech.lapsa.insurance.esbd.entities.SubjectPersonEntity;
 
 public final class TestConstants {
     // Branch
-    public static final long INVALID_BRANCH_ID = 999999999L;
+    public static final int INVALID_BRANCH_ID = 999999999;
 
     // CancelationReason
     public static final int INVALID_CANCELATION_REASON_ID = 999999999;
 
     // City
-    public static final long INVALID_CITY_ID = 999999999L;
+    public static final int INVALID_CITY_ID = 999999999;
 
     // CompanyActivityKind
-    public static final long INVALID_COMPANY_ACTIVITY_KIND_ID = 999999999L;
+    public static final int INVALID_COMPANY_ACTIVITY_KIND_ID = 999999999;
 
     // CountryRegion
     public static final int INVALID_COUNTRY_REGION_ID = 999999999;
@@ -31,7 +32,7 @@ public final class TestConstants {
     public static final int INVALID_IDENTITY_CARD_TYPE_ID = 999999999;
 
     // InsuranceCompany
-    public static final long INVALID_INSURANCE_COMPANY_ID = 999999999L;
+    public static final int INVALID_INSURANCE_COMPANY_ID = 999999999;
 
     // MaritalStatus
     public static final int INVALID_MARITAL_STATUS_ID = 999999999;
@@ -40,7 +41,7 @@ public final class TestConstants {
     public static final int INVALID_SEX_ID = 999999999;
 
     // User
-    public static final long INVALID_USER_ID = 999999999L;
+    public static final int INVALID_USER_ID = 999999999;
 
     // InsuranceAgeExpirience
     public static final int INVALID_INSURANCE_AGE_EXPIRIENCE_ID = 999999999;
@@ -51,27 +52,27 @@ public final class TestConstants {
     public static final String INVALID_INSURANCE_CLASS_TYPE_CODE = "QQ";
 
     // Subject
-    public static final long INVALID_SUBJECT_PERSON_ID = 999999999L;
-    public static final long VALID_SUBJECT_PERSON_ID = 14132412; // Evsyukovs ID
+    public static final int INVALID_SUBJECT_PERSON_ID = 999999999;
+    public static final int VALID_SUBJECT_PERSON_ID = 14132412; // Evsyukovs ID
     public static final InsuranceClassType VALID_CLASS_TYPE_FOR_CLIENT = InsuranceClassType.CLASS_13; // Evsyukovs
 													      // Class
     public static final int INVALID_VEHICLE_AGE_CLASS_ID = 999999999;
     public static final int INVALID_VEHICLE_CLASS_ID = 999999999;
 
     // VehicleManufacturer
-    public static final long VALID_VEHICLE_MANUFACTURER_ID = 45755L; // INFINTI
+    public static final int VALID_VEHICLE_MANUFACTURER_ID = 45755; // INFINTI
 								     // FX35
     public static final String VALID_VEHICLE_MANUFACTURER_NAME = "INFINTI FX35";
-    public static final long ININVALID_VEHICLE_MANUFACTURER_ID = 999999999L;
+    public static final int ININVALID_VEHICLE_MANUFACTURER_ID = 999999999;
     public static final String INVALID_VEHICLE_MANUFACTURER_NAME = "QQQ";
 
     // VehicleModel
-    public static final long VALID_VEHICLE_MODEL_ID = 143827L;
+    public static final int VALID_VEHICLE_MODEL_ID = 143827;
     public static final String VALID_VEHICLE_MODEL_NAME = "INFINITI FX35";
-    public static final long ININVALID_VEHICLE_MODEL_ID = 999999999L;
+    public static final int ININVALID_VEHICLE_MODEL_ID = 999999999;
 
     // Subject
-    public static final long[] VALID_SUBJECT_IDS = new long[] { 1, 100 };
+    public static final int[] VALID_SUBJECT_IDS = new int[] { 1, 100 };
     public static final String[] VALID_SUBJECT_ID_NUMBERS = new String[] {
 	    "930840000071",
 	    "581114350286" };
@@ -82,23 +83,23 @@ public final class TestConstants {
 	    SubjectCompanyEntity.class,
 	    SubjectPersonEntity.class };
 
-    public static final long INVALID_SUBJECT_ID = -1L;
+    public static final int INVALID_SUBJECT_ID = -1;
     public static final String INVALID_SUBJECT_ID_NUMBER = "6666";
 
     // SubjectCompany
-    public static final long[] VALID_SUBJECT_COMPANY_IDS = new long[] { 1, 2 };
-    public static final long[] INVALID_SUBJECT_COMPANY_IDS = new long[] { 100, -1 };
+    public static final int[] VALID_SUBJECT_COMPANY_IDS = new int[] { 1, 2 };
+    public static final int[] INVALID_SUBJECT_COMPANY_IDS = new int[] { 100, -1 };
     public static final String[] VALID_SUBJECT_COMPANY_BINS = new String[] { "930840000071" };
     public static final String[] INVALID_SUBJECT_COMPANY_BINS = new String[] { "581114350286", "ZZZZ" };
 
     // SubjectPerson
-    public static final long[] VALID_SUBJECT_PERSON_IDS = new long[] { 100, 14132412 };
-    public static final long[] INVALID_SUBJECT_PERSON_IDS = new long[] { 1, 2, -1 };
+    public static final int[] VALID_SUBJECT_PERSON_IDS = new int[] { 100, 14132412 };
+    public static final int[] INVALID_SUBJECT_PERSON_IDS = new int[] { 1, 2, -1 };
     public static final String[] VALID_SUBJECT_PERSON_IINS = new String[] { "581114350286", "870622300359" };
     public static final String[] INVALID_SUBJECT_PERSON_IINS = new String[] { "930840000071", "ZZZZ" };
 
     // Policy
-    public static final Long INVALID_POLICY_ID = 1L;
+    public static final Integer INVALID_POLICY_ID = 1;
     public static final String INVALID_POLICY_NUMBER = "ZZZ";
 
     // KZCity
